Analysing Data

GPMethylation.run_gpregression — Function

run_gpregression(samplemetafile, betafile, m=50; verbose=true)

Run analysis for model selection between constant, linear or non-linear (Mat52) kernels

Input:

samplefmetafile - File with sample meta

betafile - RData for with probe beta.

outdir - Directory to store results, with following contents

gpstats.tsv - Tab separated file with marginal likelihoods and optimised kernel parameters for all kernels

gpmeanvar.rdata - RData file containing GP mean and variance for each probe

m = 50 - Number of points to calculate GP mean and variance at between min and max Age.

GPMethylation.saverdata — Method

saverdata(probes, samples, GPT, outdir; verbose=true)

Save gp mean and var in RData format.

GPMethylation.gpstats — Method

gpstats(probes, GPT, beta)

Writes tab-separated file with gp stats.