Gaussian Process Regression

GPMethylation.gpreg_all_threads — Method

gpreg_all_threads(t, beta, st, gpreg, ; nt = 16, bt = 1, n = size(beta, 2))

Wrapper to run a Gaussian Process regression function gp_reg ∈ {gp_reg_matern52, gp_reg_const, gp_reg_linear} with multiple threads.

Keyword argument nt = 16 specifies number of threads.

Allows balance of threads to run regression on independent probes in parallel vs threads used by BLAS library for parallelising linear algebra used in regression of a single probe.

Currently, recommend setting BLAS threads bt = 1 and parallelising over multiple probes.

GPMethylation.gpmodels — Method

gpmodels(t, y, st)

Runs GP regression for constant, linear and Matern52 kernels.

GPMethylation.gpreg — Function

gpreg(t, y, st, kernel, kernparams=(1.0, 1.0), logNoise=log(std(y)))

Generic GP stationary zero mean regression with Gaussian likelihood. Params:

t vector of time points (in this case ages)

y vector of signal values (in this case prob methylation)

st vector of time points to evaluate GP mean and var

kernel GP kernel with inital params e.g. Mat52Iso

logσn log(σn)

Returns: Named tuple (gp=gp, μ=μ, v=v)

GPMethylation.gpreg_matern52 — Function

gpreg_matern52(t, y, st, kernparams=(1.0, 1.0) , log(std(y))))

GP regression with Mat52Iso(ℓ, σf) kernel.

kernparms = (logℓ, log(σf))

ℓ - initial lengthscale

σf - initial function variance

GPMethylation.gpreg_const — Function

gpreg_const(t, y, st, kernparams=(1.0) , log(std(y))))

GP regression with Const(σf).

kernparms = (log(σf))

σf - initial function variance

GPMethylation.gpreg_linear — Function

gpreg_linear(t, y, st, kernparams=(1.0) , log(std(y))))

GP regression with Const(σf) + LinIso(ℓ) kernel.

kernparms = (logℓ, log(σf))

ℓ - initial lengthscale

σf - initial function variance