Loading Data

GPMethylation.loaddata — Method

loaddata(file)

Loads methylation data.

file is a RData file containing a matrix named betas.dasen of size (n x m) for n samples and m probes.

Sample names and probe names taken from colnames and rownames respectively.

GPMethylation.loadsamplemeta — Method

loadsamplemeta(file)

Loads sample meta, a tab or commma separated file that must contain the following first threecolumns (in any order):

Sample	Sex	Age	Filter

Sample_1	F	1.0	1

Sample_2	M	2.0	0

Sample_3	M	10.0	1

Sample unique string identifier

Sex {M, F} or missing

Age floating point age

Filter optional column of {1, 0}s to exclude samples, when Filter = 0 sample is excluded.

Sample should match colnames of RData file.

GPMethylation.loadprobemeta — Method

loadprobemeta(file)

Loads probe meta, a tab or commma separated file that must contain the following columns (in any order):

Probe

cg00000029

cg00000103

cg00000109

Can be used to override the rownames in RData file.

GPMethylation.loadmeta — Method

loadmeta(file, mandatorycols, label)

Helper function used to load metafiles

GPMethylation.loadall — Method

loadall(samplemetafile, betafile)

Loads samplefmetafile and RData file betafile ensures samples details correspond and filters samples