Analysing Data

GPMethylation.run_gpregression — Function

run_gpregression(samplemetafile, betafile, m=50; verbose=true)

Run analysis for model selection between constant, linear or non-linear (Mat52) kernels

Input:

samplefmetafile - File with sample meta

betafile - RData for with probe beta.

outdir - Directory to store results, with following contents

gpstats.tsv - Tab separated file with marginal likelihoods and optimised kernel parameters for all kernels

gpmeanvar.rdata - RData file containing GP mean and variance for each probe

m = 50 - Number of points to calculate GP mean and variance at between min and max Age.

GPMethylation.saverdata — Method

saverdata(probes, samples, μ, v, outdir; verbose=true)

Save gp mean and var in RData format.

GPMethylation.gpstats — Method

gpstats(probes, dfgp, beta, gpμ)

Writes tab-separated file with gp stats.

GPMethylation.loadresults — Method

loadresults(outdir)

Loads gpstats and GP mean and var

GPMethylation.loadresults_rdata — Method

loadresults_rdata(file)

Loads gpmean and gpvar results from RData file back into Julia.